requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
pandas>=2.0.0
openai>=1.0.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0
